prep excersises are done
